Move histogram code to separate file.
